+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## 0.1.0
 
 * Support FFI 1.0.0
+* Add crop operation
+* Add rotation operation
+* Change API to `bitmap.apply(Operation())`
+* Add batch operations
+* Update plugin to support Flutter 1.20.0
 
 ## 0.0.1